+++ app/controllers/courts_controller.rb
@@ -4,20 +4,21 @@ extend Authorization
   authorize [ :destroy, :index, :create, :update, :edit], "master"
 
   def create
-    @court = Court.new( params[ :court])
-    if @court.save
-      flash[ :success] = t( "court.created", court: @court.name)
+    court = Court.new
+    court.massign params[ :court], :name, :link
+    if court.save
+      flash[ :success] = t( "court.created", court: court.name) if court
       redirect_to courts_path
     else
       @courts = Court.all
-      @new_court = Court.new
+      @new_court = court
       render :index
     end
   end
 
   def index
     @courts = Court.all
-    @new_court = Court.new :link => nil
+    @new_court = Court.new
   end
 
   def edit
@@ -26,19 +27,26 @@ extend Authorization
 
   def update
     @court = Court.find params[ :id]
-    if @court.update_attributes( params[ :court])
-      flash.now[ :success] = t( "court.changed")
-      redirect_to courts_path
+    if @court
+      @court.massign params[ :court], :name, :link
+      if @court.save
+        flash.now[ :success] = t( "court.changed")
+        redirect_to courts_path
+      else
+        render :edit
+      end
     else
-      render :edit
+      redirect_to courts_path
     end
   end
 
   def destroy
     destroyed = Court.find params[ :id]
-    destroyed_inspect = destroyed.inspect
-    destroyed.destroy
-    flash[ :success] = t( "court.destroyed", court: destroyed_inspect)
+    if destroyed
+      destroyed_inspect = destroyed.inspect
+      destroyed.destroy
+      flash[ :success] = t( "court.destroyed", court: destroyed_inspect)
+    end
     redirect_to courts_path
   end
 end
