+++ app/controllers/users_controller.rb
@@ -3,18 +3,19 @@ extend Authorization
 
   authorize [ :index, :show], [ "normal", "admin", "master"]
   authorize [ :edit, :update],
-            [ "disabled", "normal", "admin", "master"] do |id, user|
-    user.id == id.to_i || user.master? ||
-      (user.admin? && User.find( id).court == user.court)
+            [ "disabled", "normal", "admin", "master"] do |params, user|
+    user.id == params[ :id].to_i || user.master? ||
+      (user.admin? && User.find( params[ :id]).court == user.court)
   end
   authorize [ :disable, :enable, :promote, :destroy],
-            [ "admin", "master"] do |id, user|
-    if user.id == id.to_i
+            [ "admin", "master"] do |params, user|
+    if user.id == params[ :id].to_i
       false
     else
-      affected_user = User.find id
-      (user.master? && !affected_user.master?) ||
-        (!affected_user.admin? && affected_user.court == user.court)
+      affected_user = User.find params[ :id]
+      affected_user &&
+        ( (user.master? && !affected_user.master?
+          ) || (!affected_user.admin? && affected_user.court == user.court))
     end
   end
 
@@ -25,14 +26,15 @@ extend Authorization
   end
 
   def create
-    court_id = params[ :user].delete( :court).to_i
-    @user = User.new( params[ :user])
-    @user.court_id = court_id
-    if @user.save
-      log_in @user
-      flash[ :success] = t( "user.created", name: @user.name)
+    user = User.new
+    user.massign params[ :user],
+                 :court, :email, :name, :password, :password_confirmation
+    if user.save
+      log_in user
+      flash[ :success] = t( "user.created", name: user.name)
       redirect_to root_path
     else
+      @user = user
       @courts = Court.all
       render "new"
     end
@@ -48,23 +50,20 @@ extend Authorization
     end
   end
 
-  def show
+  def show; show_or_edit end
+  def edit; show_or_edit end
+  def show_or_edit
     @user = User.find params[ :id]
+    redirect_to users_path unless @user
   end
-
-  def edit
-    @user = User.find params[ :id]
-  end
+  private :show_or_edit
 
   def update
     @user = User.find params[ :id]
-    case params[ :commit]
-    when t( "booking.morning.book")     then update_book_do( :morning)
-    when t( "booking.afternoon.book")   then update_book_do( :afternoon)
-    when t( "booking.morning.unbook")   then update_unbook_do( :morning)
-    when t( "booking.afternoon.unbook") then update_unbook_do( :afternoon)
-    else
-      if @user.update_attributes( params[ :user])
+    if @user
+      @user.massign params[ :user],
+                    :email, :name, :password, :password_confirmation
+      if @user.save
         if current_user? @user
           log_in @user  # because remember_token has been reset
           flash.now[ :success] = t( "user.changed.message")
@@ -76,59 +75,42 @@ extend Authorization
       else
         render 'edit'
       end
+    else
+      redirect_to users_path
     end
   end
 
-  def disable
-    update_role_do( "disabled")
-  end
-  def enable
-    update_role_do( "normal")
-  end
-  def promote
-    update_role_do( "admin")
-  end
-
-  def destroy
-    destroyed = User.find params[ :id]
-    destroyed_inspect = destroyed.inspect
-    destroyed.destroy
-    flash[ :success] = t( "user.destroyed", user: destroyed_inspect)
-    redirect_to users_path
-  end
-
-  def update_book_do( session)
-    @user.bookings.create! :court_day_id => params[ :court_day],
-                           :session => session
-  rescue ActiveRecord::RecordInvalid => e
-    flash[ :error] = e.message
-  ensure
-    back_to_court_days
-  end
-  private :update_book_do
-
-  def update_unbook_do( session)
-    @user.booked?( CourtDay.find( params[ :court_day]), session).destroy
-    back_to_court_days
-  end
-  private :update_unbook_do
-
+  def disable; update_role_do( "disabled") end
+  def enable;  update_role_do( "normal")   end
+  def promote; update_role_do( "admin")    end
   def update_role_do( role)
     @user = User.find params[ :id]
-    old_role = @user.role
-    if User.valid_role?( role) && @user.update_attribute( :role, role)
-      flash[ :success] = t( "user.role.changed",
+    if @user
+      old_role = @user.role
+      if User.valid_role?( role) && @user.update_attribute( :role, role)
+        flash[ :success] = t( "user.role.changed",
+                              name: @user.name,
+                              from: t( "user.role.#{ old_role}"),
+                              to: t( "user.role.#{ role}"))
+      else
+        flash[ :error] = t( "user.role.change_fail",
                             name: @user.name,
                             from: t( "user.role.#{ old_role}"),
                             to: t( "user.role.#{ role}"))
-    else
-      flash[ :error] = t( "user.role.change_fail",
-                          name: @user.name,
-                          from: t( "user.role.#{ old_role}"),
-                          to: t( "user.role.#{ role}"))
+      end
     end
     redirect_to users_path
   end
   private :update_role_do
+
+  def destroy
+    destroyed = User.find params[ :id]
+    if destroyed
+      destroyed_inspect = destroyed.inspect
+      destroyed.destroy
+      flash[ :success] = t( "user.destroyed", user: destroyed_inspect)
+    end
+    redirect_to users_path
+  end
 end
